requirements.txt
# Core Framework
fastapi==0.116.1
uvicorn==0.35.0
pydantic==2.11.7
python-dotenv==1.1.1

# Celery for Async Task Processing
celery[redis]==5.3.4
redis>=5.0.0

# Firebase/Firestore
firebase_admin==7.1.0

# Neo4j Knowledge Graph
neo4j==5.28.2

# AI/ML - Groq API
groq==0.31.1

# LangChain Framework
langchain==0.3.27
langchain_core==0.3.75
langchain_community==0.3.29
langchain_groq==0.3.7
langchain_ollama==0.3.7
langchain_openai==0.3.32

# LlamaIndex Framework
llama_index==0.14.0

# LightRAG Framework
lightrag==0.1.0b6

# Vector Search - FAISS
faiss-gpu-cu12==1.11.0

# Embeddings & Reranking
sentence_transformers==4.1.0
ragatouille==0.0.9.post2
rank_bm25==0.2.2

# PyTorch with CUDA 12.8
torch==2.8.0.dev20250516+cu128
transformers==4.51.3

# Document Processing
pdfplumber==0.11.6
pymupdf>=1.23.0
unstructured==0.19.0
beautifulsoup4==4.12.3

# Excel Processing
openpyxl==3.1.5
pandas==2.3.2

# Word Processing
python-docx>=1.1.0

# PDF Table Extraction
camelot-py[cv]>=0.11.0
opencv-python>=4.8.0

# Image Processing
Pillow==11.3.0
pytesseract>=0.3.10

# Legal NLP Extraction
lexnlp==1.9.0

# Natural Language Processing
nltk==3.9.1
tiktoken==0.11.0

# Mathematical & Symbolic Computation
asteval==1.0.6
sympy==1.14.0

# HTTP & Async
requests==2.32.5
httpx==0.27.0

# Rate Limiting
slowapi==0.1.9

# Retry Logic
tenacity==9.1.2

# Fuzzy String Matching
rapidfuzz==3.14.1

# RAG Evaluation
ragas==0.3.3

# BEIR Benchmarking
beir==2.2.0

# Numerical Computing
numpy==2.3.3

# Telegram Bot (optional)
python-telegram-bot>=21.0

# Additional utilities
werkzeug>=3.0.0
sse-starlette>=1.8.0
# Note: bitsandbytes removed - requires C compiler and CUDA dev setup

